requirements.txt
# ============================================================
# NotificAI — Dependências
# ============================================================

# --- Ingestão de Dados ---
pdfplumber>=0.10.0          # Extração nativa de PDF
PyMuPDF>=1.23.0             # Extração PDF (fallback) — pacote: fitz
pdf2image>=1.16.0           # Conversão PDF → imagem para OCR
pytesseract>=0.3.10         # OCR (requer Tesseract-OCR instalado no sistema)
Pillow>=10.0.0              # Manipulação de imagens
python-docx>=1.1.0          # Extração de documentos Word
openpyxl>=3.1.0             # Leitura de .xlsx
pandas>=2.0.0               # Leitura de .xlsx, .csv, e manipulação de dados
chardet>=5.0.0              # Detecção de encoding para .txt

# --- NLP ---
spacy>=3.7.0                # NLP / NER (modelo pt_core_news_sm instalado via:
                            #   python -m spacy download pt_core_news_sm)
scikit-learn>=1.4.0         # Classificador ML (TF-IDF + LogisticRegression)

# --- API ---
fastapi>=0.111.0            # Framework REST
uvicorn[standard]>=0.29.0   # Servidor ASGI
python-multipart>=0.0.9     # Upload de arquivos via FastAPI

# --- Exportação e Integração ---
gspread>=6.0.0              # Google Sheets (exportação legada)
google-auth>=2.29.0         # Auth Google

# --- Dev & Testes ---
pytest>=8.0.0
httpx>=0.27.0               # Client HTTP para testes da API FastAPI

# ============================================================
# Instalação do sistema (Ubuntu/Debian):
#   sudo apt install tesseract-ocr tesseract-ocr-por poppler-utils
#
# Instalar modelo spaCy PT:
#   python -m spacy download pt_core_news_sm
# ============================================================
